разобьет main на компоненты
